Code Cleanup & Fixed JSON Error
reverie.py wasn't automatically creating the /movement directory if it did not yet exist when trying to write a new timestep to it. This is now fixed. The appropriate actions are now written to the JSON file for that timestep, however we have a new issue. System hangs.

Extraneous ChatGPT plugin code has been removed from run_gpt_prompt.py and replaced with the relevant v2 template methods.